requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv>=1.0.1,<2.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
requests==2.31.0
pytest>=8.2.0,<9.0.0
httpx[socks]==0.25.1
elasticsearch==7.17.0
openai>=1.0.0,<2.0.0
langchain>=0.0.267
langchain-openai>=0.0.2